refactor router types and includes
