+++ include/boost/http.hpp
@@ -44,7 +44,7 @@
 #include <boost/http/rfc/upgrade_rule.hpp>
 
 #include <boost/http/server/cors.hpp>
+#include <boost/http/server/route_handler.hpp>
 #include <boost/http/server/router.hpp>
-#include <boost/http/server/basic_router.hpp>
 
 #endif
