+++ src/server/router.cpp
@@ -7,7 +7,7 @@
 // Official repository: https://github.com/cppalliance/http
 //
 
-#include <boost/http/server/router.hpp>
+#include <boost/http/server/route_handler.hpp>
 
 namespace boost {
 namespace http {
