+++ test/unit/server/router_types.cpp
@@ -8,8 +8,8 @@
 //
 
 // Test that header file is self-contained.
+#include <boost/http/server/route_handler.hpp>
 #include <boost/http/server/router.hpp>
-#include <boost/http/server/basic_router.hpp>
 
 #include "test_suite.hpp"
